Switch to explicit registration of application listener in plugin.xml
Fixes #248
